LayoutTests/imported/w3c/web-platform-tests/workers/semantics/multiple-workers/008-expected.txt

CONSOLE MESSAGE: line 4: ReferenceError: Can't find variable: SharedWorker

FAIL messagechannel in shared worker Can't find variable: SharedWorker
